Corrected pointer type for adc1_buf

+++ components/vds/Core/Src/main.c
@@ -146,7 +146,7 @@ int main(void)
   /* USER CODE BEGIN 2 */
   HAL_CAN_Start(&hcan1);
   HAL_CAN_Start(&hcan2);
-  HAL_ADC_Start_DMA(&hadc1, (uint32_t *)adc1_buf, ADC1_BUF_LENGTH); // Start ADC1 in DMA mode
+  HAL_ADC_Start_DMA(&hadc1, (uint16_t *)adc1_buf, ADC1_BUF_LENGTH); // Start ADC1 in DMA mode
   HAL_TIM_Base_Start(&htim3);
 
   /* USER CODE END 2 */
